+++ src/promptflow/tests/sdk_cli_test/conftest.py
@@ -11,7 +11,6 @@ from pytest_mock import MockerFixture
 from sqlalchemy import create_engine
 
 from promptflow import PFClient
-from promptflow._core.openai_injector import inject_openai_api
 from promptflow._sdk._configuration import Configuration
 from promptflow._sdk._constants import EXPERIMENT_CREATED_ON_INDEX_NAME, EXPERIMENT_TABLE_NAME, LOCAL_MGMT_DB_PATH
 from promptflow._sdk._serving.app import create_app as create_serving_app
@@ -20,6 +19,7 @@ from promptflow._sdk.entities._connection import CustomConnection, _Connection
 from promptflow._utils.utils import is_in_ci_pipeline
 from promptflow.executor._line_execution_process_pool import _process_wrapper
 from promptflow.executor._process_manager import create_spawned_fork_process_manager
+from promptflow.tracing._openai_injector import inject_openai_api
 
 from .recording_utilities import (
     RecordStorage,
@@ -293,15 +293,15 @@ def setup_recording_injection_if_enabled():
             "promptflow._core.tool.tool": mocked_tool,
             "promptflow._internal.tool": mocked_tool,
             "promptflow.tool": mocked_tool,
-            "promptflow._core.openai_injector.inject_sync": inject_sync_with_recording,
-            "promptflow._core.openai_injector.inject_async": inject_async_with_recording,
+            "promptflow.tracing._openai_injector.inject_sync": inject_sync_with_recording,
+            "promptflow.tracing._openai_injector.inject_async": inject_async_with_recording,
         }
         start_patches(patch_targets)
 
     if is_live() and is_in_ci_pipeline():
         patch_targets = {
-            "promptflow._core.openai_injector.inject_sync": inject_sync_with_recording,
-            "promptflow._core.openai_injector.inject_async": inject_async_with_recording,
+            "promptflow.tracing._openai_injector.inject_sync": inject_sync_with_recording,
+            "promptflow.tracing._openai_injector.inject_async": inject_async_with_recording,
         }
         start_patches(patch_targets)
